Fix multi-select on structured view
