+++ src/javascript/SelectorTypes/Picker/PickerDialog/RightPanel/ContentLayout/ContentTable/ContentTable.jsx
@@ -34,6 +34,7 @@ import classes from './ContentTable.scss';
 import {registry} from '@jahia/ui-extender';
 import {useFieldContext} from '~/contexts/FieldContext';
 import {configPropType} from '~/SelectorTypes/Picker/configs/configPropType';
+import {flattenTree} from '~/SelectorTypes/Picker/Picker2.utils';
 
 const reduxActions = {
     onPreviewSelectAction: () => ({}),
@@ -80,11 +81,13 @@ export const ContentTable = ({rows, isContentNotFound, totalCount, isLoading, ca
     };
 
     const columns = useMemo(() => {
+        const flattenRows = isStructured ? flattenTree(rows) : rows;
         if (pickerConfig?.pickerTable?.columns) {
             const columns = pickerConfig.pickerTable.columns.map(c => (typeof c === 'string') ? allColumnData.find(col => col.id === c) : c);
-            if (field.multiple && rows.some(r => r.isSelectable) && !columns.find(c => c.id === 'selection')) {
+            if (field.multiple && flattenRows.some(r => r.isSelectable) && !columns.find(c => c.id === 'selection')) {
                 columns.splice((columns[0].id === 'publicationStatus') ? 1 : 0, 0, allColumnData[1]);
             }
+
             return columns;
         }
 
@@ -92,8 +95,8 @@ export const ContentTable = ({rows, isContentNotFound, totalCount, isLoading, ca
             // Do not include type column if media mode
             .filter(c => Constants.mode.MEDIA !== mode || c.id !== 'type')
             // Do not include selection if multiple selection is not enabled or if there are no selectable types
-            .filter(c => (field.multiple && rows.some(r => r.isSelectable)) || c.id !== SELECTION_COLUMN_ID);
-    }, [mode, field.multiple, pickerConfig, rows]);
+            .filter(c => (field.multiple && flattenRows.some(r => r.isSelectable)) || c.id !== SELECTION_COLUMN_ID);
+    }, [mode, field.multiple, pickerConfig, rows, isStructured]);
     const {
         getTableProps,
         getTableBodyProps,
